Added ownership to displayed Items, Refactoring and Exception catching

+++ src/routers/event.py
@@ -1,25 +1,64 @@
+"""
+Router class for events, providing all features needed to create, view, join and delete events.
+"""
 from fastapi import *
 from typing import List
 from src.util.token import *
-from ..util.database import init_db
-from ..models.event import Event, AccountEvent
-from ..schemas.event import *
+from src.util.database import init_db
+from src.models.event import Event, AccountEvent
+from src.schemas.event import *
+
 
 router = APIRouter()
 
 
 @router.get("/", response_model=List[RespondEvent])
 def get_all(db: Session = Depends(init_db)):
+    """
+    Get all events registered in DB
+    :param db: DB do browse with
+    :return: All events in DB
+    """
     return db.query(Event).all()
 
 
 @router.get("/{id}", response_model=RespondEvent)
-def get_by_id(id: int, db: Session = Depends(init_db)):
-    return db.query(Event).filter(Event.eventID == id).first()
+def get_by_id(id: int, token:str = Header(None), db: Session = Depends(init_db)):
+    """
+    Get a specific event.
+    :param id: Unique ID of the event.
+    :param db: DB to browse
+    :return: Event identified by the ID
+    """
+    if db.query(Event).filter(Event.eventID == id).first() is None:
+        raise HTTPException(status_code=404)
+    ownership = False
+    data = db.query(Event).filter(Event.eventID == id).first()
+    if token is not None:
+        user = read_token(token, db)
+        if data.creator == user:
+            ownership = True
+    response = RespondEvent(
+    eventID=data.eventID,
+    creator=data.creator,
+    eventname=data.eventname,
+    location=data.location,
+    appointment=data.appointment,
+    maxAttendants=data.maxAttendants,
+    ownership=ownership
+    )
+    return response
 
 
 @router.post("/")
 def add_event(request: RequestEvent, token: str = Header(None), db: Session = Depends(init_db)):
+    """
+    Add an event to the DB.
+    :param request: Request body to create event
+    :param token: Token to identify logged in user
+    :param db: DB to browse
+    :return: OK if success
+    """
     if read_token(token, db) is None:
         raise HTTPException(status_code=401)
     new_event = Event(
@@ -27,14 +66,25 @@ def add_event(request: RequestEvent, token: str = Header(None), db: Session = De
         eventname=request.eventname,
         location=request.location,
         appointment=request.appointment,
-        maxAttendants=request.maxAttendants
+        maxAttendants=request.maxAttendants,
+        description=request.description
     )
     db.add(new_event)
     db.commit()
+    return {"response": "ok"}
 
 
 @router.post("/join/{id}")
 def join_event(id: int, token: str = Header(None), db: Session = Depends(init_db)):
+    """
+    Enter a specific event identified by unique ID
+    :param id: Unique event ID
+    :param token: Token to identify user, check if logged in
+    :param db: DB to browse
+    :return: OK if success
+    """
+    if db.query(Event).filter(Event.eventID == str(id)).first() is None:
+        raise HTTPException(status_code=404)
     if read_token(token, db) is None:
         raise HTTPException(status_code=401)
     new_participant = AccountEvent(
@@ -43,16 +93,41 @@ def join_event(id: int, token: str = Header(None), db: Session = Depends(init_db
     )
     db.add(new_participant)
     db.commit()
+    return {"response": "ok"}
 
 
 @router.get("/participants/{id}")
 def get_participants(id: int, db: Session = Depends(init_db)):
+    """
+    Get joined users of a specific event
+    :param id: Unique ID of event
+    :param db: DB to browse
+    :return: List of all joined accounts
+    """
+    if db.query(Event).filter(Event.eventID == str(id)).first() is None:
+        raise HTTPException(status_code=404)
     return db.execute("select * from account where email in (select email from account_event where event = " + str(id) + ")").fetchall()
 
 
 @router.delete("/{id}")
-def delete_event(id: int, db: Session = Depends(init_db)):
+def delete_event(id: int, token: str = Header(None), db: Session = Depends(init_db)):
+    """
+    Delete an event.
+    You have to be logged in as the owner of the event for this.
+    :param id: Unique ID of Event
+    :param token: Token to identify user
+    :param db: DB to browse
+    :return: OK if success
+    """
+    user = read_token(token, db)
+    if db.query(Event).filter(Event.eventID == str(id)).first() is None:
+        raise HTTPException(status_code=404)
+    if user is None:
+        raise HTTPException(status_code=401)
+    if user != db.query(Event.creator).filter(Event.eventID == str(id)).first()[0]:
+        raise HTTPException(status_code=401)
     db.execute("DELETE FROM account_event WHERE event = " + str(id))
     db.commit()
     db.execute("DELETE FROM event WHERE eventID = " + str(id))
     db.commit()
+    return {"response": "ok"}
